Fix HoldableWeights not have its collision turned back on if it lands in a bowl.
This was happening because bowls were constantly unparenting and reparenting weights.  Each time a node changes parents, it skips its _PhysicsProcess step for a little bit.  This resulted in _PhysicsProcess never getting to run for weights while they're in a bowl, meaning ThrowableParentKinematic never had a chance to turn its collision back on.